Implement equality for different currencies
A Money instance is now equal to another if the instance it's being
compared to can be converted to the same currency. Otherwise it will
throw a ConversionError as normal.

I'm hesitant to add this, it was in the instructions for the exercise
but I would probably advocate for not adding this feature. It is a bit
too magic, and I don't think that any of these two behaviours feel
natural:

- Raising an exception from #== (current behaviour, fails faster)
- Returning false if it cannot be converted, but true if it can. This
  would likely be very confusing at some point.

I'll add it for now, just to mimic the example behaviour in the README
as close as possible but I think I'd want to revert this commit in the
future.