Implement reverse conversion rates
The `require 'money_boy'` in converter.rb is very innocent looking, but
provokes *very* strange behaviour when not there. The tests would pass
when run all together, but fail when the converter_spec was run
separately, the problem is that #<=> rescues the error and the failure
says something like "expected 55.5USD, got 55.5USD", it completely
breaks equality but the actual exception is nowhere to be seen. Very
frustrating to get this bad feedback from a simple name error.